Posts --refactoring & added langs text content to form
