+++ vite-project/src/components/features/PostCard.jsx
@@ -1,4 +1,4 @@
-import {avatarUrl, dateFormat, imageUrl} from "../../app/functions.js";
+import {avatarUrl, dateFormat, imageUrl} from "../../../app/functions.js";
 import React from "react";
 import {IoEyeOutline} from "react-icons/io5";
 import {useTranslation} from "react-i18next";
@@ -11,7 +11,7 @@ export default function PostCard({post}) {
             <div className="relative w-full">
                 <img
                     src={imageUrl(post.cover.medium.url)}
-                    alt={post[locale].title}
+                    alt={post[locale]?.title}
                     className="h-56 sm:h-72 w-full rounded-2xl bg-gray-100 object-cover"
                 />
                 <div className="absolute inset-0 rounded-2xl ring-1 ring-inset ring-gray-900/10"/>
@@ -25,17 +25,17 @@ export default function PostCard({post}) {
                         href={`/posts/${post.id}`}
                         className="relative z-10 rounded-full bg-gray-50 px-3 py-1.5 font-medium text-gray-600 hover:bg-gray-100"
                     >
-                        {post[locale].category}
+                        {post[locale]?.category}
                     </a>
                 </div>
                 <div className="group relative">
                     <h3 className="mt-3 text-lg font-semibold leading-6 text-gray-900 group-hover:text-gray-600">
                         <a href={`/posts/${post.id}`}>
                             <span className="absolute inset-0"/>
-                            {post[locale].title}
+                            {post[locale]?.title}
                         </a>
                     </h3>
-                    <p className="mt-5 line-clamp-3 text-sm leading-6 text-gray-600">{post[locale].desc}</p>
+                    <p className="mt-5 line-clamp-3 text-sm leading-6 text-gray-600">{post[locale]?.desc}</p>
                 </div>
                 <div className="relative mt-8 mb-4 flex items-center gap-x-4">
                     <img
